Remove: Mocks were removed
