+++ src/utils/statistics.js
@@ -1,4 +1,4 @@
-import {EmojiValue} from './constants';
+import {EmojiValue} from './constants.js';
 
 export const getLegendWithEmoji = (name) => {
   const emoji = EmojiValue[name.split(`-`)[0].toUpperCase()];
